Delete ref to no-longer-existing Design Guide.
Design Guide is still on the public web at https://www.learnersguild.org/learning-os-software/design-guide. However, it was deleted from this repo on 11 August 2017. I surmise this is because it was exclusively chat-interface-focused and the LOS is making web interfaces more prominent. The Design Guide was not replaced or rewritten for web-oriented design, so there is (apparently) nothing internal to reference. An external reference could be inserted, but that would seem to require a high-level decision on what design principles or theory to adopt.