Incorporate sole convention into best practices
Outline of this section differed from that of the previous section. The previous one had 2 subsections, but both were linked to from inside the main file. This section had 1 subsection, NOT linked to from inside the main file. This could cause a reader to mistakenly assume that he had already read the convention subsection. Since the convention file had only 1 section, i.e. file structure, it seemed more user-friendly to append that to the existing content of the main (README) file. One might argue that a convention is not a best practice, but if so then a file on conventions should not be a subsection of the best-practices section. This change implies deleting the convention file and its entry in the summary file.